refact: Move fetch logic to its own hook
